requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-filter==21.1
django-import-export==2.8.0
django-rest-swagger==2.2.0
drf-yasg==1.20.0